% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brvm-allindex.R
\name{BRVM.index}
\alias{BRVM.index}
\title{BRVM Available Indexes}
\usage{
BRVM.index()
}
\value{
A character
}
\description{
It receives no argument and returns BRVM's indexes informations.
}
\examples{
\dontrun{BRVM.index()}

}
\seealso{
Other BRVM: 
\code{\link{BRVM_cap}()},
\code{\link{BRVM_company_cap}()},
\code{\link{BRVM_company_rank}()},
\code{\link{BRVM_company_url}()},
\code{\link{BRVM_index_stock}()},
\code{\link{BRVM_index}()},
\code{\link{BRVM_market_activity}()},
\code{\link{BRVM_plot}()},
\code{\link{BRVM_stock_market}()},
\code{\link{BRVM_ticker_desc}()},
\code{\link{BRVM_traded_val}()},
\code{\link{company_cap}()},
\code{\link{company_country}()},
\code{\link{company_nbrank}()},
\code{\link{company_traded_val}()},
\code{\link{normality_test}()},
\code{\link{stationarity_test}()}
}
\author{
Koffi Frederic SESSIE
}
\concept{BRVM}
