% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util-brvm-sectors.R
\name{BRVM_available_sectors}
\alias{BRVM_available_sectors}
\title{BRVM Available Sectors}
\usage{
BRVM_available_sectors()
}
\value{
A tibble
}
\description{
Returns a tibble of available sectors for BRVM.
}
\details{
This function will return a vector tibble of available sectors.
}
\examples{

BRVM_available_sectors()

}
\seealso{
\url{"https://www.richbourse.com/common/apprendre/liste-societes"}

\url{"https://www.richbourse.com/common/apprendre/liste-societes?page=2"}

\url{"https://www.richbourse.com/common/apprendre/liste-societes?page=3"}

Other Utility: 
\code{\link{BRVM_available_tickers}()}
}
\author{
Koffi Frederic SESSIE

Oudouss Diakité Abdoul

Steven P. Sanderson II, MPH
}
\concept{Utility}
