% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util-brvm-tickers.R
\name{BRVM_available_tickers}
\alias{BRVM_available_tickers}
\title{BRVM Available Tickers}
\usage{
BRVM_available_tickers()
}
\value{
A tibble
}
\description{
Returns a tibble of available Tickers for BRVM.
}
\details{
This function will return a vector tibble of available Tickers.
}
\examples{

BRVM_available_tickers()

}
\seealso{
\url{"https://www.brvm.org/en/cours-actions/"}

Other Utility: 
\code{\link{BRVM_available_sectors}()}
}
\author{
Koffi Frederic SESSIE

Oudouss Diakité Abdoul

Steven P. Sanderson II, MPH
}
\concept{Utility}
