% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brvm_bonds.R
\name{BRVM_Bonds}
\alias{BRVM_Bonds}
\title{Get BRVM Bonds info}
\usage{
BRVM_Bonds(input = "All")
}
\arguments{
\item{input}{Character string. One of "All", "State", "Private", "Regional". Returns a dataframe of bonds from BRVM. The function also uses the separate function to split the "Last payment (Date/Value)" column into "Date" and "Value" columns.}
}
\value{
dataframe
}
\description{
Get BRVM Bonds info
}
\examples{
\donttest{
library(tidyr)
library(rvest)
all_bonds = BRVM_Bonds("All")
head(all_bonds)

st_bonds = BRVM_Bonds("State")
head(st_bonds)

prv_bonds = BRVM_Bonds("Private")
head(prv_bonds)

reg_bonds = BRVM_Bonds("Regional")
head(reg_bonds)

}
}
