% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brvm-company_cap.R
\name{BRVM_company_cap}
\alias{BRVM_company_cap}
\title{BRVM companies capitalization value - To get the capitalization value data of all companies listed on BRVM stock exchange}
\usage{
BRVM_company_cap()
}
\value{
A tibble
}
\description{
This function will get the capitalization value data of all companies listed on the BVRM exchange.
The function takes no parameter
}
\examples{
\dontrun{BRVM_company_cap()}

}
\seealso{
Other BRVM: 
\code{\link{BRVM.index}()},
\code{\link{BRVM_cap}()},
\code{\link{BRVM_company_rank}()},
\code{\link{BRVM_company_url}()},
\code{\link{BRVM_index_stock}()},
\code{\link{BRVM_index}()},
\code{\link{BRVM_market_activity}()},
\code{\link{BRVM_plot}()},
\code{\link{BRVM_stock_market}()},
\code{\link{BRVM_ticker_desc}()},
\code{\link{BRVM_traded_val}()},
\code{\link{company_cap}()},
\code{\link{company_country}()},
\code{\link{company_nbrank}()},
\code{\link{company_traded_val}()},
\code{\link{normality_test}()},
\code{\link{stationarity_test}()}
}
\author{
Koffi Frederic SESSIE
}
\concept{BRVM}
