% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brvm__company_url.R
\name{BRVM_company_url}
\alias{BRVM_company_url}
\title{BRVM company url}
\usage{
BRVM_company_url(company)
}
\arguments{
\item{company}{is the name of the Ticker}
}
\value{
URL
}
\description{
It receives the ticker name and return an URL
}
\examples{
\dontrun{BRVM_company_url("BICC")}

}
\seealso{
Other BRVM: 
\code{\link{BRVM.index}()},
\code{\link{BRVM_cap}()},
\code{\link{BRVM_company_cap}()},
\code{\link{BRVM_company_rank}()},
\code{\link{BRVM_index_stock}()},
\code{\link{BRVM_index}()},
\code{\link{BRVM_market_activity}()},
\code{\link{BRVM_plot}()},
\code{\link{BRVM_stock_market}()},
\code{\link{BRVM_ticker_desc}()},
\code{\link{BRVM_traded_val}()},
\code{\link{company_cap}()},
\code{\link{company_country}()},
\code{\link{company_nbrank}()},
\code{\link{company_traded_val}()},
\code{\link{normality_test}()},
\code{\link{stationarity_test}()}
}
\author{
Koffi Frederic SESSIE
}
\concept{BRVM}
