% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brvm-top-flop.R
\name{BRVM_rank}
\alias{BRVM_rank}
\title{BRVM_rank - Get top or flop data of BRVM stock exchange}
\usage{
BRVM_rank(top_or_flop, n = 10)
}
\arguments{
\item{top_or_flop}{Choose between "top" or "flop"}

\item{n}{is the number of companies in the classification}
}
\value{
"tbl_df"     "tbl"        "data.frame"
}
\description{
It receives "top" or "flop" and a number 'n' and
returns table of companies classification
}
\examples{
\donttest{
BRVM_rank("top", 15)
BRVM_rank("flop", 5)
}
}
\author{
Koffi Frederic SESSIE

Oudouss Diakite Abdoul

Steven P. Sanderson II, MPH
}
