% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brvm-bysector.R
\name{BRVM_bySector}
\alias{BRVM_bySector}
\title{BRVM By Sector}
\usage{
BRVM_bySector(.sectors = NULL)
}
\arguments{
\item{.sectors}{A vector of sectors you wish to have returned.}

\item{.turn_off_warnings}{A boolean value of TRUE or FALSE. Should global warnings
stay on or off. The default is off, after the function finishes running it will
turn global warnings back on.}
}
\value{
A tibble
}
\description{
This function will take in the name of sector(s) and returns data for companies belonging to that/those sector(s)
}
\details{
This function will take in a vector of sectors provided by the end
user and return a tibble.
}
\examples{
\dontrun{
sectors <- c(
  "Industry", "Administration", "PUblic service", "AUtr", "FINANCE",
  "distribution", 25
)
BRVM_sector(.sectors = sectors)
BRVM_sector(.sectors = "All")
BRVM_sector(.sectors = c("Other", "Agriculture"))
}

}
\seealso{
\url{"https://www.richbourse.com/common/apprendre/liste-societes"}

\url{"https://www.richbourse.com/common/apprendre/liste-societes?page=2"}

\url{"https://www.richbourse.com/common/apprendre/liste-societes?page=3"}

Other Sector: 
\code{\link{BRVM_sector}()}

Other Data Retrieval: 
\code{\link{BRVM_cap}()},
\code{\link{BRVM_company_rank}()},
\code{\link{BRVM_direction}()},
\code{\link{BRVM_get1}()},
\code{\link{BRVM_get}()},
\code{\link{BRVM_index_stock}()},
\code{\link{BRVM_index}()},
\code{\link{BRVM_market_activity}()},
\code{\link{BRVM_plot}()},
\code{\link{BRVM_rank}()},
\code{\link{BRVM_sector}()},
\code{\link{BRVM_stock_market}()},
\code{\link{BRVM_ticker_desc}()},
\code{\link{company_cap}()},
\code{\link{company_country}()},
\code{\link{company_nbrank}()},
\code{\link{company_traded_val}()}
}
\author{
Koffi Frederic SESSIE

Oudouss Diakité Abdoul

Steven P. Sanderson II, MPH
}
\concept{Data Retrieval}
\concept{Sector}
