% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brvm_company_info.R
\name{BRVM_company_info}
\alias{BRVM_company_info}
\title{BRVM Get info about a ticker beta, RSI, Closing, Valorisation, etc.}
\usage{
BRVM_company_info(ticker)
}
\arguments{
\item{ticker}{The ticker of a company}
}
\value{
A tibble
}
\description{
It receives the ticker of a company or index listed on the BRVM stock exchange,
Turn to  upper case the input by using \code{toupper()} and returns informations about the company's RSI, Beta, Closing price, etc. .
}
\examples{
BRVM_company_info("BOAS")
BRVM_company_info("BoaM")
BRVM_company_info("BRVMAG")


}
\seealso{
\url{https://www.sikafinance.com}
}
\author{
Koffi Frederic SESSIE
}
