% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-brvm-direction.R
\name{BRVM_direction}
\alias{BRVM_direction}
\title{BRVM Rank}
\usage{
BRVM_direction(.up_or_down = "Up")
}
\arguments{
\item{.up_or_down}{This is a character string set to "Up" It can either be
'Up' or 'Down'.}
}
\value{
A tibble
}
\description{

}
\details{
This function will get any n results for the top or flop ranking
of the BRVM tickers.
}
\examples{
BRVM_direction("Up")

}
\seealso{
Other Ranking: 
\code{\link{BRVM_rank}()}

Other Data Retrieval: 
\code{\link{BRVM_get}()},
\code{\link{BRVM_index}()},
\code{\link{BRVM_rank}()},
\code{\link{BRVM_sector}()}
}
\author{
Koffi Frederic SESSIE

Oudouss Diakité Abdoul

Steven P. Sanderson II, MPH
}
\concept{Data Retrieval}
\concept{Ranking}
