% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brvm-up-down.R, R/get-brvm-direction.R
\name{BRVM_direction}
\alias{BRVM_direction}
\title{BRVM Rank}
\usage{
BRVM_direction(.up_or_down = "Up", .turn_off_warnings = TRUE)

BRVM_direction(.up_or_down = "Up", .turn_off_warnings = TRUE)
}
\arguments{
\item{.up_or_down}{This is a character string set to "Up" It can either be
'Up' or 'Down'.}
}
\value{
A tibble

A tibble
}
\description{
This function will take in 'Up' or 'Down' and returns respectively n results for the top or flop ranking
of the BRVM tickers.
}
\details{
This function will get any n results for the top or flop ranking
of the BRVM tickers.

This function will get any n results for the top or flop ranking
of the BRVM tickers.
}
\examples{
\dontrun{
BRVM_direction("Up")
}

BRVM_direction("Up")

}
\seealso{
Other Ranking: 
\code{\link{BRVM_rank}()}

Other Ranking: 
\code{\link{BRVM_rank}()}

Other Data Retrieval: 
\code{\link{BRVM_bySector}()},
\code{\link{BRVM_cap}()},
\code{\link{BRVM_company_rank}()},
\code{\link{BRVM_get1}()},
\code{\link{BRVM_get}()},
\code{\link{BRVM_index_stock}()},
\code{\link{BRVM_index}()},
\code{\link{BRVM_market_activity}()},
\code{\link{BRVM_plot}()},
\code{\link{BRVM_rank}()},
\code{\link{BRVM_sector}()},
\code{\link{BRVM_stock_market}()},
\code{\link{BRVM_ticker_desc}()},
\code{\link{company_cap}()},
\code{\link{company_country}()},
\code{\link{company_nbrank}()},
\code{\link{company_traded_val}()}
}
\author{
Koffi Frederic SESSIE

Oudouss Diakité Abdoul

Steven P. Sanderson II, MPH
}
\concept{Data Retrieval}
\concept{Ranking}
\concept{Ticker Data Retrieval}
