% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-bvrm-ticker-data.R
\name{BRVM_get}
\alias{BRVM_get}
\title{BRVM Get - Get Ticker Data}
\usage{
BRVM_get(.symbol, .from = NULL, .to = NULL)
}
\arguments{
\item{.symbol}{A vector of symbols, like: c("BICC","XOM","SlbC")}

\item{.from}{A quoted start date, ie. "2020-01-01" or "2020/01/01". The data
must be in ymd format "YYYY-MM-DD" or "YYYY/MM/DD".}

\item{.to}{A quoted end date, ie. "2022-01-31" or "2022/01/31". The data must
be in ymd format "YYYY-MM-DD" or "YYYY/MM/DD"}
}
\value{
A tibble
}
\description{
This function will get data from the Rich Bourse exchange.
}
\details{
This function will get data from the Rich Bourse exchange. The function
takes in a single parameter of \code{.symbol} The function will auto-format the
tickers you input into all upper case by using \code{toupper()} The function will
next make sure that the ticker passed is inside of a google spreadsheet of
allowed tickers.
}
\examples{
symbols <- c("BiCc","XOM","SlbC")
data_tbl <- BRVM_get(.symbol = symbols)
data_tbl

}
\seealso{
\url{https://www.richbourse.com}

Other Ticker Data Retrieval: 
\code{\link{BRVM_direction}()},
\code{\link{BRVM_index}()},
\code{\link{BRVM_rank}()}
}
\author{
Koffi Frederic SESSIE

Oudouss Diakité Abdoul

Steven P. Sanderson II, MPH
}
\concept{Richbourse}
\concept{Ticker Data Retrieval}
