% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-brvm-index.R
\name{BRVM_index}
\alias{BRVM_index}
\title{BRVM Get - Get Ticker Data}
\usage{
BRVM_index()
}
\value{
A tibble
}
\description{
This function will get data from \url{www.brvm.org}.
}
\details{
This function will get index data from \url{www.brvm.org}
}
\examples{
BRVM_index()

}
\seealso{
\url{https://www.brvm.org/en/indices/status/200}

Other Ticker Data Retrieval: 
\code{\link{BRVM_direction}()},
\code{\link{BRVM_get}()},
\code{\link{BRVM_rank}()}
}
\author{
Koffi Frederic SESSIE

Oudouss Diakité Abdoul

Steven P. Sanderson II, MPH
}
\concept{BRVM}
\concept{Ticker Data Retrieval}
