% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brvm-top-flop.R, R/get-brvm-rank.R
\name{BRVM_rank}
\alias{BRVM_rank}
\title{BRVM_rank - Get top or flop data of BRVM stock exchange}
\usage{
BRVM_rank(.top_or_flop = "Top", .n = 10, .turn_off_warnings = TRUE)

BRVM_rank(.top_or_flop = "Top", .n = 10, .turn_off_warnings = TRUE)
}
\arguments{
\item{.top_or_flop}{This is a character string set to "Top". It can either
be "Top" or "Flop".}

\item{.n}{How many records to return}

\item{top_or_flop}{Choose between "top" or "flop"}

\item{n}{is the number of companies in the classification}
}
\value{
"tbl_df"     "tbl"        "data.frame"

A tibble
}
\description{
It receives "top" or "flop" and a number 'n' and
returns table of companies classification
}
\details{
This function will get any n results for the top or flop ranking
of the BRVM tickers.
}
\examples{
\dontrun{
BRVM_rank("top", 15)
BRVM_rank("flop", 5)
}

BRVM_rank("Top", 1)
BRVM_rank("blahblah", 1)
BRVM_rank("Flop", 1)

}
\seealso{
Other Ranking: 
\code{\link{BRVM_direction}()}

Other Data Retrieval: 
\code{\link{BRVM_bySector}()},
\code{\link{BRVM_cap}()},
\code{\link{BRVM_company_rank}()},
\code{\link{BRVM_direction}()},
\code{\link{BRVM_get1}()},
\code{\link{BRVM_get}()},
\code{\link{BRVM_index_stock}()},
\code{\link{BRVM_index}()},
\code{\link{BRVM_market_activity}()},
\code{\link{BRVM_plot}()},
\code{\link{BRVM_sector}()},
\code{\link{BRVM_stock_market}()},
\code{\link{BRVM_ticker_desc}()},
\code{\link{company_cap}()},
\code{\link{company_country}()},
\code{\link{company_nbrank}()},
\code{\link{company_traded_val}()}
}
\author{
Koffi Frederic SESSIE

Oudouss Diakité Abdoul

Steven P. Sanderson II, MPH
}
\concept{Data Retrieval}
\concept{Ranking}
