% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-brvm-rank.R
\name{BRVM_rank}
\alias{BRVM_rank}
\title{BRVM Rank}
\usage{
BRVM_rank(.top_or_flop = "Top", .n = 10, .turn_off_warnings = TRUE)
}
\arguments{
\item{.top_or_flop}{This is a character string set to "Top". It can either
be "Top" or "Flop".}

\item{.n}{How many records to return}
}
\value{
A tibble
}
\description{

}
\details{
This function will get any n results for the top or flop ranking
of the BRVM tickers.
}
\examples{

BRVM_rank("Top", 1)
BRVM_rank("blahblah", 1)
BRVM_rank("Flop", 1)

}
\seealso{
Other Ranking: 
\code{\link{BRVM_direction}()}

Other Data Retrieval: 
\code{\link{BRVM_direction}()},
\code{\link{BRVM_get}()},
\code{\link{BRVM_index}()},
\code{\link{BRVM_sector}()}
}
\author{
Koffi Frederic SESSIE

Oudouss Diakité Abdoul

Steven P. Sanderson II, MPH
}
\concept{Data Retrieval}
\concept{Ranking}
