% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-brvm-sector.R
\name{BRVM_sector}
\alias{BRVM_sector}
\title{BRVM By Sector}
\usage{
BRVM_sector(.sectors = "All")
}
\arguments{
\item{.sectors}{A vector of sectors you wish to have returned.}
}
\value{
A tibble
}
\description{

}
\details{
This function will take in a vector of sectors provided by the end
user and return a tibble.
}
\examples{

sectors <- c("Industry","Administration","PUblic service", "AUtr", "FINANCE",
            "distribution", 25)
BRVM_sector(.sectors = sectors)
BRVM_sector(.sectors = "All")
BRVM_sector(.sectors = c("Other", "Agriculture"))

}
\seealso{
\url{"https://www.richbourse.com/common/apprendre/liste-societes"}

\url{"https://www.richbourse.com/common/apprendre/liste-societes?page=2"}

\url{"https://www.richbourse.com/common/apprendre/liste-societes?page=3"}
}
\author{
Koffi Frederic SESSIE

Oudouss Diakité Abdoul

Steven P. Sanderson II, MPH
}
\concept{Sector}
