% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brvm-ticker-desc.R
\name{BRVM_ticker_desc}
\alias{BRVM_ticker_desc}
\title{BRVM ticker description - BRVM ticker information}
\usage{
BRVM_ticker_desc()
}
\value{
A tibble
}
\description{
It receives no argument and returns informations about BRVM exchange companies.
}
\details{
This function gives many informations about each ticker, his full name, sector and his country
}
\examples{

BRVM_ticker_desc()


}
\seealso{
\url{https://www.brvm.org/en/cours-actions/0}

\code{BRVM_tickers()}

Other Data Retrieval: 
\code{\link{BRVM_bySector}()},
\code{\link{BRVM_cap}()},
\code{\link{BRVM_company_rank}()},
\code{\link{BRVM_get1}()},
\code{\link{BRVM_get}()},
\code{\link{BRVM_index_stock}()},
\code{\link{BRVM_index}()},
\code{\link{BRVM_market_activity}()},
\code{\link{BRVM_plot}()},
\code{\link{BRVM_stock_market}()},
\code{\link{company_cap}()},
\code{\link{company_country}()},
\code{\link{company_nbrank}()},
\code{\link{company_sector}()},
\code{\link{company_traded_val}()}

Other Richbourse: 
\code{\link{BRVM_get}()}

Other BRVM: 
\code{\link{BRVM.index}()},
\code{\link{BRVM_cap}()},
\code{\link{BRVM_company_cap}()},
\code{\link{BRVM_company_rank}()},
\code{\link{BRVM_company_url}()},
\code{\link{BRVM_index_stock}()},
\code{\link{BRVM_index}()},
\code{\link{BRVM_market_activity}()},
\code{\link{BRVM_plot}()},
\code{\link{BRVM_stock_market}()},
\code{\link{BRVM_traded_val}()},
\code{\link{company_cap}()},
\code{\link{company_country}()},
\code{\link{company_nbrank}()},
\code{\link{company_sector}()},
\code{\link{company_traded_val}()},
\code{\link{normality_test}()},
\code{\link{stationarity_test}()}
}
\author{
Koffi Frederic SESSIE

Oudouss Diakite Abdoul

Steven P. Sanderson II, MPH
}
\concept{BRVM}
\concept{Data Retrieval}
\concept{Richbourse}
