% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brvm-ticker.R
\name{BRVM_tickers}
\alias{BRVM_tickers}
\title{BRVM Tickers - Information about listed companies on BRVM Stock exchange}
\usage{
BRVM_tickers()
}
\value{
A tibble
}
\description{
It receives no argument and returns two informations about BRVM exchange companies (ticker and company name).
}
\examples{
\donttest{
BRVM_tickers()
ticks <- BRVM_tickers()
dput(ticks$Ticker) ## Returns the name of all tickers
}
}
\seealso{
\code{BRVM_ticker_desc()}
}
\author{
Koffi Frederic SESSIE

Oudouss Diakite Abdoul

Steven P. Sanderson II, MPH
}
